Renamed "rest" dir to "http"
