+++ rondb/src/main/java/site/ycsb/db/http/ds/MyHttpClientAsync.java
@@ -18,7 +18,7 @@
 /**
  * YCSB binding for <a href="https://rondb.com/">RonDB</a>.
  */
-package site.ycsb.db.rest.ds;
+package site.ycsb.db.http.ds;
 
 import org.apache.http.HttpResponse;
 import org.apache.http.client.methods.HttpRequestBase;
@@ -30,7 +30,8 @@ import org.apache.http.impl.nio.reactor.DefaultConnectingIOReactor;
 import org.apache.http.nio.reactor.ConnectingIOReactor;
 import org.apache.http.nio.reactor.IOReactorException;
 import org.apache.http.util.EntityUtils;
-import site.ycsb.db.rest.MyHttpException;
+
+import site.ycsb.db.http.MyHttpException;
 
 import java.util.concurrent.Future;
