+++ rondb/src/main/java/site/ycsb/db/http/ds/PKResponse.java
@@ -18,7 +18,7 @@
 /**
  * YCSB binding for <a href="https://rondb.com/">RonDB</a>.
  */
-package site.ycsb.db.rest.ds;
+package site.ycsb.db.http.ds;
 
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
