+++ rondb/src/main/java/site/ycsb/db/http/ds/package-info.java
@@ -18,5 +18,5 @@
 /**
  * YCSB binding for <a href="https://rondb.com/">RonDB</a>.
  */
-package site.ycsb.db.rest.ds;
+package site.ycsb.db.http.ds;
